for #887, move getTableNamesFromDefaultDataSource() to parent class
