+++ sharding-jdbc/src/main/java/io/shardingsphere/core/jdbc/metadata/dialect/H2ShardingMetaDataHandler.java
@@ -24,7 +24,6 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -55,21 +54,6 @@ public final class H2ShardingMetaDataHandler extends ShardingMetaDataHandler {
             try (ResultSet resultSet = statement.getResultSet()) {
                 while (resultSet.next()) {
                     result.add(new ColumnMetaData(resultSet.getString("FIELD"), resultSet.getString("TYPE"), resultSet.getString("KEY")));
-                }
-            }
-            return result;
-        }
-    }
-    
-    @Override
-    public Collection<String> getTableNamesFromDefaultDataSource() throws SQLException {
-        Collection<String> result = new LinkedList<>();
-        try (Connection connection = getDataSource().getConnection();
-             Statement statement = connection.createStatement()) {
-            statement.executeQuery("show tables;");
-            try (ResultSet resultSet = statement.getResultSet()) {
-                while (resultSet.next()) {
-                    result.add(resultSet.getString(1));
                 }
             }
             return result;
