+++ sharding-jdbc/src/main/java/io/shardingsphere/core/jdbc/metadata/dialect/MySQLShardingMetaDataHandler.java
@@ -24,7 +24,6 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -58,21 +57,6 @@ public final class MySQLShardingMetaDataHandler extends ShardingMetaDataHandler 
             try (ResultSet resultSet = statement.getResultSet()) {
                 while (resultSet.next()) {
                     result.add(new ColumnMetaData(resultSet.getString("Field"), resultSet.getString("Type"), resultSet.getString("Key")));
-                }
-            }
-            return result;
-        }
-    }
-    
-    @Override
-    public Collection<String> getTableNamesFromDefaultDataSource() throws SQLException {
-        Collection<String> result = new LinkedList<>();
-        try (Connection connection = getDataSource().getConnection();
-             Statement statement = connection.createStatement()) {
-            statement.execute("show tables;");
-            try (ResultSet resultSet = statement.getResultSet()) {
-                while (resultSet.next()) {
-                    result.add(resultSet.getString(1));
                 }
             }
             return result;
